+++ src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
-export default function Home() {
+export default function Dashboard() {
   return (
-    <main>
+    <main className="container">
       <h1 className="text-3xl font-bold underline">Hello world!</h1>
     </main>
   );