+++ src/app/layout.tsx
@@ -4,8 +4,6 @@ import { Inter } from "next/font/google";
 import "../styles/globals.css";
 import "@rainbow-me/rainbowkit/styles.css";
 import { ThemeProvider } from "@/contexts/theme-provider";
-import { Web3ModalProvider } from "@/contexts/web3-modal-provider";
-import { Header } from "@/components/shared/header";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -22,12 +20,7 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={inter.className}>
-        <ThemeProvider>
-          <Web3ModalProvider>
-            <Header />
-            {children}
-          </Web3ModalProvider>
-        </ThemeProvider>
+        <ThemeProvider>{children}</ThemeProvider>
       </body>
     </html>
   );