+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# [0.4.0](https://github.com/tianbuyung/web3-hub/compare/v0.3.0...v0.4.0) (2024-08-04)
+
+
+### Features
+
+* **web3:** add RainbowKit, Wagmi, Viem, React Query ([605d012](https://github.com/tianbuyung/web3-hub/commit/605d012eefb66d93d28720cb2c86b7fa7a815de5))
+* **web3:** setup wagmi config, update next.js and create connect wallet modal ([b40af83](https://github.com/tianbuyung/web3-hub/commit/b40af832212c360ee3de1cfcc5489b9542be7322))
+
 # 0.2.0 (2024-07-29)
 
 ### Bug Fixes